Move LoadingPlaceholder and OptionalValue components to a common place
