+++ src/components/UI/Display/MAPI/clusters/ClusterList/ClusterListItemMainInfo.tsx
@@ -5,7 +5,7 @@ import * as React from 'react';
 import styled from 'styled-components';
 import { Dot } from 'styles';
 import KubernetesVersionLabel from 'UI/Display/Cluster/KubernetesVersionLabel';
-import OptionalValue from 'UI/Display/MAPI/clusters/ClusterDetail/OptionalValue';
+import OptionalValue from 'UI/Display/OptionalValue/OptionalValue';
 
 const StyledDot = styled(Dot)`
   padding: 0;
